Ported individual project pages to /project/[project_id]
